pyproject.toml
[project]
name = "simple-agent-for-mirador-with-claude-2025-08-05"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.61.0",
    "dotenv>=0.9.9",
    "mcp>=1.12.3",
    "strands-agents>=1.3.0",
    "streamlit>=1.48.0",
]
